Refactor: move page create/update logic into hook
