+++ src/components/PageSettingsModal/PageSettingsModal.jsx
@@ -84,23 +84,10 @@ export const PageSettingsModal = ({
   /** ******************************** */
 
   const onSubmit = (data) => {
-    const trimmedNameData = {
-      ...data,
-      title: data.title.trim(),
-    }
     return onProceed({
-      frontMatter:
-        pageData && pageData.content
-          ? {
-              ...pageData.content.frontMatter,
-              ...trimmedNameData,
-            }
-          : trimmedNameData,
-      sha: pageData?.sha || "",
-      pageBody: pageData?.content?.pageBody || "",
-      newFileName: resourceRoomName
-        ? `${trimmedNameData.date}-${trimmedNameData.layout}-${trimmedNameData.title}.md`
-        : `${trimmedNameData.title}.md`,
+      pageData,
+      data,
+      resourceRoomName,
     })
   }
